Added show_test.rb and shows_test.rb with some refactoring of the Shows data model
Part of the work done here cleans up the data model for the Shows object. Instead of being an
array available across all show objects which could cause weird bugs down the road, shows are
now an array local to the instantiated object. Also, I added a method to allow for adding
shows on the fly.